Only allow rotations in RigidBoneSystem
The bone transforms produced from the rigid bone system now consist only
of rotations around a center point, where the center point has a scaling
and translation of the original bone baked into it.

Constraining the types of transformed allowed in the RigidBoneSystem
will make it easier to implement IK systems.