Use isIkable bone attribute to determine which bone was grabbed
Previously I was using the "face groups" defined in the Daz asset along
with a hardcoded list of non-IKable face groups.

The new technique produces slightly better looking results because the
bone grabbed is now always the one which will most affect the closest
face whereas the Daz face group map was only an approximation.