Fix getting stuck minimized
This was a tricky one. Since introducing the AsyncFramePrepare, the
Viewer sometimes gets stuck in the minimized state if you minimize it.

It appears what's happening is that the SemaphoreSlim I'm using for
synchronization will pump messages if it waits, but it somehow does this
incorrectly and restore message is dropped.

The solution is to use a SpinWait to wait for the prepared frame instead
of a Semaphore. This is probably the right thing to do anyway since
timing is critical and the prepared frame we usually be ready
immediately (rendering usually takes longer than frame preparation).