Rotate about whole figure about the root bone instead of the hips
Both the "hips" bone and the root bone rotate the whole figure (though
about different center points) so it only make sense for the IK solver
to change out of them. Previously, the IK solver rotated about the hips
and left the root rotation untouched, but I'm changing it to rotate
about the root and leave the hips untouched. The reason is that I want
to take advantage of the specially handling that the root bone gets when
calculating IK deltas (all axes are treated equally on the root rotation
whereas the other bones treat their twist axes differently).