Prevent spurious WasClicked events when activating controllers
The WasClicked event is defined as a button become unpressed when it was
previously pressed. This can happen spuriously if a button is pressed
when the controller becomes active; then releasing the button causes a
click event.

This is happening during the Save workflow. The OpenVR keyboard closes
and the controller reactivates while the trackpad is still pressed. Then
when a trackpad is released, a spurious Click event is trigger.

To prevent this, I'm adding a new field groundState to
ControllerStateTracker that tracks which buttons have at some point
being unpressed since the controller was activated. Click events will
now only trigger if the button has at some point been inactive since the
controller was activated.