Add a false depth pass after rending one-sided volumes
The false depth pass pushes the depth back so that two-sided surfaces
(e.g., clothing) will tend to be drawn in front off the volumes even if
they geometrically intersect (which could happen after applying some
morphs).

Right now the false depth is a constant 2 centimetres behind the true
depth. I intended to refine this so that it takes into account the shape
of the figure.