Add support for triangular faces in meshes
All faces are still stored as four indices but the last index is
duplicated for triangles. The rest of the code recognizes these
"degenerate quads" and handles them correctly. In particular:
1) The OpenSubdivFacade handles triangles in the control topology and
produces triangles in the refined topology with correct stencils.
2) Anywhere I convert a quad to two triangles for rendering now produces
a single triangle from degenerate quads.

I didn't update ScatteringFormFactorCalculator and
BoneAttributesCalculator to handle triangular faces because these
classes only operate on the main figure mesh which contains no
triangles.