Split figure surfaces into opaque and transparent per face
Previously, an entire surface had to be tagged as either opaque and
transparent; now, transparency is set for an individual control face.
The result is that manny faces previously in a transparent layer are now
in an opaque layer. This is both improves performance (because
transparent surfaces are slower) and visual quality (because some
comprises between rendering speed and quality are made for transparent
surfaces).