Fix Eva face variant bug
The "Eva 7 Face 01.duf" file has a bug that causes the face turn pink
(it happens in Daz Studio as well). I could make an asset patch that
fixes the problem, but since "Face 01" is the default face material,
it's easier just to inherit the default from "Eva 7 All Maps.duf" which
doesn't have the bug.