Split content into content packs
1) The import configuration files have been split up into
subdirectories.
2) At import time, the importer runs through each import configuration
subdirectory and writes the associated figures and outfits to the
"work/content/<content-pack-name>" subdirectory.
3) At pack time, the packer loops through each subdirectory under
"work/content" and packs its contents into a file under
"work/packed-content".
4) At run time, the viewer scans the "content" directory (which can be
repointed via the command line) and unions any the subdirectories and
packed archives it contains. This union is used as the root archive
directory for the app.

Currently, materials and shapes must be in the same content pack as the
figure they belong to. This means the "core" content pack, containing
the G3F figure, is still very large. In a subsequent check-in, I'll
allow materials and shapes to go in separate content packs.