Store channel values instead of deltas in ActorModel
I want to store actual channel values instead of deltas in the actor
model so that the actual values appear in the menu system and persisted
scenes.

This hasn't matter until now because all channels that appeared in the
menu system had initial values of zero. But I wan't to add a scale
channel with an initial value of 1.