Add Locked property to Channels
Locked channels cannot be changed (for example, extra rotation channels
on Twist/Swing bones or channels used as intermediate values for
formulas). Previously the Visible property was serving double-duty both
to prevent channels from appearing in the UI and prevent them from being
set internal. But I want to use Visible exclusively for the former and
Locked for the latter.