Hide modifiers coming from add-on products
Some character products, particularly non-"Daz original" ones, add extra
morphs. I don't want these morphs to be visible for a couple of reasons:
1) They clutter up to UI. I had previously been handling this by
hardcoding a list of channel paths to hide in the viewer menu system,
but hiding on a product-wide basis is cleaner.
2) They interfere with selecting channels for the occlusion system. In
particular, if an add-on product included a visible morph that depended
on a core morph, then the core morph was being dropped from the
occlusion system.